+++ mail/mail-uniffi/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
   - Exposed `delete_contact` functionality.
+  - Exposed `watch_contact_list` with new `ContactsLiveQueryCallback` interface.
 
 ## [0.20.0] - 2024-10-31